reorganization of groups for quantum tensor
Added typetraits to detect whether a simple group is acceptable for the
group compositor cgroup.

add cgroup reference type: cgroup_ref and cgroup_cref.
These reference types are fully interoperable with the cgroup.
Their role is to allow easy manipulation of the element of an upcoming
polymorphic container type for the concrete cgroups. Elements of this
container will be manipulated through cgroup_ref or cgroup_cref objects.

otherwise a memory continuguous (and polymorphic) container is much trickier to make and
manipulate.

Further testing still necessary. It is very likely to be filled with
silly little bug.