refactored all group stuff, now conserved quantity
The fact that the implemented conserved quantity are discrete groupis more
coincidental than anything (it's alway going to be true for conservation
law associated to abelian symmetry of the system). A conserved quantity
is a label for an irreducible representation of the group, the
composition law of irreps do not follow group properties in general.