documentation targets for CMAKE
Those target are available only if CMAKE detect that the relevent
programs are available.

If everything is available, there's the plain old doxygen website as
well as a more modern (and much more customisable) sphinx website.

The sphinx website require doxygen,sphinx,breathe and the sphinx_rtd_theme.

The Documentation target will rebuilt everything it can if it detect a
change in the headers or if there's a change to the settings of doxygen
or sphinx.