Reshape tools, tensordot can do tensor product
The reshape tools are the shape_from functions.
The member shape_from slices dimensions
The free function compute the shape of the tensor product.
if A,B,C are tensors, it is meant to be used in concert with reshape_as
like so:
    btensor D = C.reshape_as(shape_from(A.shape_from(-1,2,-1),B);
A is a rank 3 tensor, we take it's first and last dimension, and all of
the dimensions of B to form a new shape.
This suppose C conserved value are  compatible with the shape so
described.
If it is not, an exception will be thrown by reshape_as