Loads of change, working toward block tensor dmrg
dmrg now return a tensor instead of a scalar for the energy, this way it is possible to use automatic differentiation through it. I do not know if this is a good idea.

new functions and struct for btensor dmrg

btensor now hold a torch option structure and can be set during construction.
btensor has a few free standing function for boolean and arithmetic operations

btensor.Conj() now also inverse all the conserved quantities ase well as conjugated  complex numbers.
When conserved quantities are involved for quantum mechanics, the hermitian conjugate must also inverse the conserved quantities. mean conj() always come with this operation. This make conversion from non-conserving to conserving much easier. btensor.conj_onlyt() does complex conjugation without changing the conserved quantities if that is necesarry.

btensor.to() to change the torch::TensorOption of a tensor.

Created a torch_shape object to manipulate shape without without having to manipulate full tensor. This make it possible to manipulate shapes in pretty much the same way for basic tensor as for full tensors.

created factory function for btensor, rand, full, ones... They all only fill the position allowed by the selection rule of the shape. They all have a _like variant that create a tensor using the shape of another.

conserved quantities must now implement a "distance" concept (and distance squared). This is useful for the generation of random networks with an overall selection rule. The distance squared must be composable, meaning that the distance square between two composed quantities is the sum of the distance squared of the components. This change trickles all the way to any_quantities.

Discovered that torch doesn't support negative steps for reverse iterations. That's a bummer, and violate their promise of the same interface as basic python arrays.

implemented linear algebra SVD and symeig for btensor.

MPO and MPS check rank always tested right side of the left side of the same tensor, that was wrong. It would have given both false positive and false negative.

btensor::check_product_compat violated its noexcept specification if there was a mismatch in block sizes, corrected.

implemented bMPT,bMPO and bMPS. the basic tensor network for btensors.
 mirror MPT,MPO and MPS.

 I made way to many change without any commit i know.